Move deck page into CubeLayout
