+++ src/cube_deck.js
@@ -1,13 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import DraftDeck from './components/DraftDeck';
-import ErrorBoundary from './components/ErrorBoundary';
+import CubeDeckPage from 'components/CubeDeckPage';
 
-const element = (
-  <ErrorBoundary className="mt-5">
-    <DraftDeck {...reactProps} />
-  </ErrorBoundary>
-);
+const element = <CubeDeckPage {...reactProps} />;
 const wrapper = document.getElementById('react-root');
 wrapper ? ReactDOM.render(element, wrapper) : false;
