+++ src/layouts/CubeLayout.js
@@ -16,7 +16,7 @@ const CubeNavItem = ({ link, activeLink, children }) => {
 
 const CubeLayout = ({ cube, cubeID, canEdit, activeLink, children }) => {
   return (
-    <CubeContextProvider initialCube={cube.cards} cubeID={cubeID} canEdit={canEdit}>
+    <CubeContextProvider initialCube={cube.cards} cubeID={cubeID}>
       <link rel="stylesheet" href="/css/autocomplete.css" />
       <ul className="cubenav nav nav-tabs nav-fill d-flex flex-column flex-sm-row pt-2">
         <div className="nav-item px-lg-4 px-3 text-sm-left text-center font-weight-boldish mt-auto mb-2">
