+++ views/cube/cube_deck.pug
@@ -1,29 +1,15 @@
-extends cube_layout
+extends ../layout
 
-block cube_toolbar
+block content
   link(rel='stylesheet' href='/css/draft.css')
 
-block cube_content
   include ../dynamic_flash
+
   #react-root
 
 block scripts
   script(src='/js/autocard.js')
   script(type='text/javascript').
-    var reactProps = {
-      oldFormat: !{oldformat},
-      drafter: !{JSON.stringify(drafter)},
-      cards: !{cards ? JSON.stringify(cards) : '[]'},
-      deck: !{deck || []},
-      description: !{JSON.stringify(description)},
-      name: !{JSON.stringify(name)},
-      sideboard: !{sideboard || []},
-      botDecks: !{JSON.stringify(bot_decks)},
-      bots: !{JSON.stringify(bots)},
-      canEdit: !{JSON.stringify(user && user._id == owner)},
-      comments: !{JSON.stringify(comments)},
-      deckid: !{JSON.stringify(deckid)},
-      userid: !{JSON.stringify(user._id)},
-    };
+    var reactProps = !{JSON.stringify(reactProps)};
   include ../react
   script(src='/js/cube_deck.bundle.js')
