+++ views/cube/cube_list.pug
@@ -11,6 +11,7 @@ block content
 
   include ../react
 
+block scripts
   script(src='/js/autocard.js', async)
   script(type='text/javascript').
     var reactProps = !{JSON.stringify(reactProps)};
